@jahorton 2024-04-02
So, what's the performance for loading a keyboard look like presently?

![image](https://github.com/keymanapp/keyman/assets/25213402/e11e8be2-570e-42f4-ae66-f79486c3f3d1)

The green box is the part that this helps optimize.

What about when we first swap to a new layer?

![image](https://github.com/keymanapp/keyman/assets/25213402/3f70ddf6-efdb-4b34-b414-86fcf84f2045)

The big "recalculate style" is needed to ensure that each key receives proper styling.  (The layer has become visible [`display: block`] for the first time, a prerequisite to getting valid style information.)  After all, it is quite possible for keyboard authors (see: `sil_cameroon_azerty`, among other keyboards) to set key-specific CSS stylesheet styling... and we can only get that when each element fully enters the DOM.   Note that once the _one_ style-recalculation pass is complete... no more reflow-related segments show up, with the layout-recalculation segment passing pretty quickly otherwise.