@jahorton 2024-04-03
Also of note:  I have done nothing to mitigate or prevent extra reflows from occurring due to predictive-text banner functionality - and it _is_ triggering some of the reflow seen in the images above - especially within the 3x context reset section.

Not shown - the layout refresh once the stylesheets finish loading.  It's not a keyboard-loading bottleneck, at least, and is pretty well optimized now.