@jahorton 2024-04-03
With regard to the predictive-text banner:

![image](https://github.com/keymanapp/keyman/assets/25213402/34d67757-2ac6-496b-abbf-7977412426b5)

Note that there is a `configureForKeyboard` function that may be used to reset things upon a keyboard change in addition to a more-standard `refreshLayout()` that can handle device rotations, etc.  We should be able to precompute the font data for standard keystrokes in order to mitigate layout reflow here.

Fortunately, this is the _one_ spot where the predictive-text banner actually polls _any_ style information on a per-keystroke basis.  Even then... we could just "store" the changes in a closure and execute them at the same time as we make changes in the main OSK's layout.

Another offender looks to be the banner's `updateFade` method, which doesn't double-check whether or not any style edits are truly needed:

https://github.com/keymanapp/keyman/blob/73b8ff2bb4c0c9217fab991dd5c6b00d10a0bc23/web/src/engine/osk/src/banner/suggestionBanner.ts#L203-L226

There's nearly always some reflow visible within the profiling associated with its portions on the timeline.