@bharanidharanj 2024-04-03
# Test Results

* **TEST_IOS_PHONE (PASSED):** Tested KeymanWeb on iPhone 13 Mobile device (iOS 17.4) and did not encounter any issues. Seems to be working fine.