@bharanidharanj 2024-04-03
# Test Results

* **TEST_ANDROID_TABLET (PASSED):** Tested using the KeymanWeb test page on an Android 11.0 Tablet / API 30 emulator and here is my observation: 1. Executed all Touch platform tests, including those focused on gestures, and confirmed their successful operation. 2. Certain tests were omitted in the Android tablet emulator due to the requirement of simultaneous key holding and long-pressing for result observation (e.g., Test_Numeric_From_Shift, Test_Delayed_Subkey, Test_Flick_During_Modipress).