@jahorton 2024-04-04
Huh, another neat thing to note...

![image](https://github.com/keymanapp/keyman/assets/25213402/41a1abdb-fcaf-439b-8e5c-6b30b8a2ef38)

The act of fetching previously-untapped cookies takes a surprising amount of time, it seems.