@jahorton 2024-04-04
After running several profiles to collect approximate time data using the in-office SM-T350 test device (Android 6.0.1), I have the following details to report:

### Keyboard load times

Keyboards used:  `sil_euro_latin`, `sil_cameroon_azerty`, `khmer_angkor`.

This data does _not_ include the time spent waiting for the keyboard stylesheets to be applied.  It only measures "time until usable."

With #11129 and #11140 in place:
- keyboard first-loads require between 1.0 and 1.2 seconds.
    - That is, when a keyboard is first loaded into the host page.
- `khmer_angkor` (with fewer layers) seems to load between 0.1 and 0.2 seconds faster than `sil_euro_latin` and `sil_cameroon_azerty`.
- _Re_loading `sil_euro_latin` (after `khmer_angkor`) tended to be about 0.1 to 0.2 seconds faster than its initial load.
    - There was negligible cost incurred by swapping lexical models, and we even got net _savings_ due to the keyboard having previously been loaded!

Without them in place:
- `sil_euro_latin`'s first load landed consistently around 1.8 seconds.
- `sil_cameroon_azerty`'s first load landed consistently around 2.25 seconds!
    - In retrospect, this was likely penalized in part by the _deactivation_ of the English lexical model.
- `khmer_angkor` locked in between 1.5 seconds and 1.6 seconds.
- _Re_loading `sil_euro_latin` after `khmer_angkor` tended to cost 2.4 seconds!
    - The increased time was due to extra context resets and related layout reflows from _deactivating_ the Khmer lexical model and _then_ activating the English lexical model.  Both steps performed context resets.

Conclusion:  
- simple keyboards tend to load about 33% faster - in 2/3 of the time - as they did before these two PRs.
- complex keyboards may load nearly twice as fast!
- Swapping between multiple keyboards, each supporting predictive text, no longer incurs a penalty.

### App-keyboard initialization

How much time is required to activate Keyman Engine for Web within the Keyman app keyboard's host page in total?  Start to usable-finish?

I happened to notice significant difference in the amount of time it took to reset for each profile between the old and new versions of the code, so I thought this number might prove interesting.  And, it does!

Default keyboard: `sil_euro_latin`.

1. Base cost:  Using the `--debug` version of KeymanWeb at the core, the base KeymanWeb script took approximately 1.0 to 1.1 seconds to load.

After that...

With #11129 and #11140 in place:

2. Between 475 to 560 ms were taken to initialize Keyman Engine for Web (empty keyboard due to async script loading)
3. Approximately 1 second was taken to load and activate `sil_euro_latin`, starting from when the keyboard's script first began to be processed.

For an approximate total average time, let's go with 2.6 seconds to initialize.  (1.1 + 0.5 + 1)

Without them in place:

2. Between 810 to 850 ms were taken to initialize Keyman Engine for Web (empty keyboard due to async script loading)
3. Approximately 2.3 to 2.4 seconds were taken to load and activate `sil_euro_latin`, starting from when the keyboard's script first began to be processed.
    - I am not immediately sure why there's extra time compared to the base load.
    - I _will_ note, though, that I tended to leave out small intermediate "tasks" unrelated to the main thing being measured in the previous section.  It's quite possible that those add up significantly. 

For an approximate total average time, let's go with 4.2 seconds to initialize.  (1.1 + 0.8 + 2.3)

2.6 / 4.2 = 0.6190%
5.2 / 2.6 ~= 1.6154.

So.., the host-page now initializes about 60% faster / in about 62% of the time it used to take!

### Repeated layer-swapping

This is the aspect most directly targeted by #11129, though #11140 also brings some improvements of its own to the table. 

My goal with this test was to test the performance of cross-layer multitapping.  Sadly, it seems that **when the profiler is active**, it somehow interferes with proper layer-changing multitap processing.  That being said, the simple act of swapping layers repeatedly in close succession still yields some remarkable insights worth discussing.

The input sequence:
- On `sil_euro_latin`, with an empty context and the SHIFT layer active...
- Triple-tap the numeric key (which should rotate through to "currency", then "symbolic")
    - But with the profiler active, just... swapped back and forth from numeric to default.
- Then single tap any _one_ standard output key.

With #11129 and #11140 in place:

- If no layer changes had previously occurred, the input sequence tended to require between 1.3 to 1.4 seconds, with occasional small gaps in the timeline.  (~95% - 98% utilization?)
- Resetting the context and layer, then repeating:  the input sequence tended to require approximately 1.1 seconds.
    - 1.35 / 1.1 = 1.227, or approximately a 22.7% speedup when revisiting previously-accessed layers.

Without these two PRs in place:

- If no layer changes had previously occurred, the input sequence tended to require between 3 to 3.6 seconds, without _any_ gaps in the timeline, indicating full 100% utilization.
    -  3.3 / 1.35 = 2.44.... or a speedup of approximately 144%!
- Resetting the context and layer, then repeating:  the input sequence tended to require between 2.5 to 2.7 seconds.
    - 2.6 / 1.1 = 2.3636..., approximately 136% faster than the old code.
    - Even in the original code, there was approximately a relative 27% speedup (3.3 / 2.6 = 1.269) when revisiting previously-accessed layers, which isn't too far from the 22.7% figure we see with the post-PR code.

### Concluding Thoughts

Note that the _many_ of the big penalties being eliminated in _this_ specific section would have only been introduced in 17.0-alpha, so it's not likely to be _that_ big of a speedup from 16.0.  The _other_ sections, though?  Fair game.

As such, I think we can safely claim a 33% speed keyboard-swapping speed increase, a 60% app-keyboard initialization-speed increase, and probably at _least_ a 33%, if not 50%, layer-swapping speed increase relative to 16.0-stable.  I understand if we'd rather get hard numbers from 16.0 rather than recent 17.0-beta code before making that claim, though.