@jahorton 2024-04-05
Here's one sample Chrome profile _after_ these changes:

[android host-page init profile.json](https://github.com/keymanapp/keyman/files/14878676/android.host-page.init.profile.json)

I would consider everything through 3563ms, and maybe the next "Task" afterward, part of host-page initialization.  The "Task" from around 1550ms to 2043ms would be the KMW initializing with an empty keyboard for the OSK (followed by layout from around 2100ms to 2237ms), with the task starting at 2242ms or so being when the actual keyboard is first loaded.

Obviously, spending that much time to load an empty keyboard, just to turn around and load a proper one, isn't ideal.  That might be something worth looking at further.  This PR does at least make both keyboard loads notably faster, but what'd be even faster is to build the OSK once (for the right keyboard) instead of twice (once for a stand-in default, then once for the right keyboard).

![zoomed-out profile timeline](https://github.com/keymanapp/keyman/assets/25213402/94049892-5ba8-426d-a6e1-d04575464f6d)